+++ CHANGELOG.md
@@ -34,7 +34,7 @@ Released on ??
 > 🦥 The lazy update
 
 - **Default ssh config path**:
-  - SSH configuration path is not `~/.ssh/config` by default
+  - SSH configuration path is now `~/.ssh/config` by default
 - **Bugfix**:
   - Fixed [Issue 126](https://github.com/veeso/termscp/issues/126)
   - Fixed [Issue 141](https://github.com/veeso/termscp/issues/141)